Rename MakeRoomAdminCommand to HijackRoomCommand
