+++ src/commands/HijackRoomCommand.ts
