+++ test/integration/commands/hijackRoomCommandTest.ts
